perf: address review - queue memory fix, extract asyncPool to utils (#385)
