+++ apps/server/src/infrastructure/api/routers/media-router.ts
@@ -6,6 +6,7 @@ import {
 } from "@solid-imager/core/domain/media/schemas";
 import { z } from "zod";
 import { MediaService } from "~/application/services/media-service";
+import { asyncPool } from "~/utils/async-pool";
 
 /**
  * Media Router Implementation
@@ -230,35 +231,3 @@ export const mediaRouter = {
 				}),
 		),
 };
-
-/**
- * Process items with a concurrency limit.
- * Returns per-item results via Promise.allSettled-style entries.
- */
-async function asyncPool<T, R = void>(
-	items: T[],
-	limit: number,
-	fn: (item: T) => Promise<R>,
-): Promise<PromiseSettledResult<R>[]> {
-	const results: PromiseSettledResult<R>[] = new Array(items.length);
-	let index = 0;
-
-	async function worker() {
-		while (true) {
-			const i = index++;
-			if (i >= items.length) break;
-			try {
-				const value = await fn(items[i]);
-				results[i] = { status: "fulfilled" as const, value };
-			} catch (reason) {
-				results[i] = { status: "rejected" as const, reason };
-			}
-		}
-	}
-
-	const workers = Array.from({ length: Math.min(limit, items.length) }, () =>
-		worker(),
-	);
-	await Promise.all(workers);
-	return results;
-}
